Extract local storage logic to its own class
